// src/interfaces/twitter_profile.ts
export interface ITwitterProfile {
  issued: boolean;
  twitter_avatar: string;
  twitter_followers_count: number;
  twitter_id: string;
  twitter_listed_count: number;
  twitter_name: string;
  twitter_tweet_count: number;
  twitter_username: string;
  twitter_verified: boolean;
  address: string;
  public_code: string;
  twitterAvatar: string;
}